feat: add rimraf, update bundle, and dynamify the ThemeProvider
