+++ package.json
@@ -15,7 +15,7 @@
     "node": ">=20.7.0"
   },
   "scripts": {
-    "bundle": "microbundle --jsx React.createElement --jsxImportSource react --compress true --format modern,cjs --visualize",
+    "bundle": "rimraf dist && microbundle --jsx React.createElement --jsxImportSource react --compress true --format modern,cjs --visualize",
     "build": "npm run bundle",
     "dev": "npm run bundle -- watch",
     "prepare": "run-s build",
@@ -69,6 +69,7 @@
     "microbundle": "^0.15.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.0.4",
+    "rimraf": "^6.0.1",
     "sass": "^1.56.2",
     "semantic-release": "^19.0.5",
     "storybook": "^8.1.11",
